test: stabilize group access privacy gate
